raspberry/requirements.txt
scapy==2.4.5
zdm-client-py==1.0.1
requests==2.28.1
python-dotenv==0.21.0